pyproject.toml
[project]
name = "cryptotest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorlog>=6.9.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
